We improve the performance as some users experience a significant slow-down since th[e last PR.](https://github.com/cowprotocol/services/pull/610) 

See discussion [here](https://cowservices.slack.com/archives/C0361CDD1FZ/p1667035528153689):

I added another test -called postgres_user_orders_performance_user_with_many_orders  - that actually tests the case of one user having a lot of orders (like 100k orders). I tried different implementations and measured the performance:

Implementations
UNION: This is implementation as on main branch right now.
OR: This is the implementation of this PR
UNION with ordering twice: This adds additional orderings to each subtable in the UNION solution. The hope was that these orderings could be used by postgres to end queries early.

In the following I document the performances of the test: postgres_user_orders_performance_many_users_with_some_orders with i in 1..100, j in 1..1000
postgres_user_orders_performance_user_with_many_orders with i in 0..1, j in 1..100k


UNION:
```
9.065873ms
test orders::tests::postgres_user_orders_performance_many_users_with_some_orders ... ok
1.269072995s
test orders::tests::postgres_user_orders_performance_user_with_many_orders ... FAILED
```

OR
```
6.597712ms
test orders::tests::postgres_user_orders_performance_many_users_with_some_orders ... ok
163.476722ms
test orders::tests::postgres_user_orders_performance_user_with_many_orders ... ok
```

UNION solution and ordering twice

```
1.566726733s
test orders::tests::postgres_user_orders_performance_user_with_many_orders ... FAILED
7.531417ms
test orders::tests::postgres_user_orders_performance_many_users_with_some_orders ... ok
```

Looking at the results, its clear that we want to the OR solution and we will receive a 9x improvement. Then the [following API request](https://api.cow.fi/xdai/api/v1/account/0x69d962A66E5BfFf0875cB7a6345bE283bC6437d2/orders) would take 2 secs, but rather 300 ms.

If you guys have other ideas, let me know.

### Test Plan