@josojo 2022-11-03
>After reading up on the first performance improving PR I'm a bit surprised about the performance of the OR solution. Didn't you original benchmarks show that it performs worse than UNION for users with few orders?

yes, it must have been an outlier or mistake. 
In order to exclude outliers in the future, I am now averaging the performance of many requests in the test.